+++ frontend/src/scenes/dashboard/dashboards/Dashboards.tsx
@@ -14,7 +14,9 @@ import { DashboardTemplatesTable } from 'scenes/dashboard/dashboards/templates/D
 import { newDashboardLogic } from 'scenes/dashboard/newDashboardLogic'
 import { SceneExport } from 'scenes/sceneTypes'
 
-import { SceneContent, SceneDivider, SceneTitleSection } from '~/layout/scenes/SceneContent'
+import { SceneContent } from '~/layout/scenes/components/SceneContent'
+import { SceneDivider } from '~/layout/scenes/components/SceneDivider'
+import { SceneTitleSection } from '~/layout/scenes/components/SceneTitleSection'
 import { dashboardsModel } from '~/models/dashboardsModel'
 import { AccessControlLevel, AccessControlResourceType } from '~/types'
