+++ frontend/src/scenes/insights/insightSceneLogic.tsx
@@ -43,13 +43,13 @@ import { parseDraftQueryFromLocalStorage, parseDraftQueryFromURL } from './utils
 const NEW_INSIGHT = 'new' as const
 export type InsightId = InsightShortId | typeof NEW_INSIGHT | null
 
-function isDashboardFilterEmpty(filter: DashboardFilter | null): boolean {
+export function isDashboardFilterEmpty(filter: DashboardFilter | null): boolean {
     return (
         !filter ||
-        (filter.date_from == null &&
-            filter.date_to == null &&
-            (filter.properties == null || (Array.isArray(filter.properties) && filter.properties.length === 0)) &&
-            filter.breakdown_filter == null)
+        (filter.date_from === null &&
+            filter.date_to === null &&
+            (filter.properties === null || (Array.isArray(filter.properties) && filter.properties.length === 0)) &&
+            filter.breakdown_filter === null)
     )
 }
 
@@ -366,7 +366,7 @@ export const insightSceneLogic = kea<insightSceneLogicType>([
             }
 
             const dashboardName = dashboardLogic.findMounted({ id: dashboard })?.values.dashboard?.name
-            const filtersOverride = searchParams['filters_override']
+            const filtersOverride = dashboardLogic.findMounted({ id: dashboard })?.values.temporaryFilters
             const variablesOverride = searchParams['variables_override']
 
             if (
