+++ frontend/src/scenes/web-analytics/tabs/marketing-analytics/frontend/components/settings/NativeExternalDataSourceConfiguration.tsx
@@ -10,7 +10,7 @@ import { cn } from 'lib/utils/css-classes'
 import { DataWarehouseSourceIcon } from 'scenes/data-warehouse/settings/DataWarehouseSourceIcon'
 import { urls } from 'scenes/urls'
 
-import { SceneSection } from '~/layout/scenes/SceneContent'
+import { SceneSection } from '~/layout/scenes/components/SceneSection'
 import { ExternalDataSource, PipelineNodeTab, PipelineStage } from '~/types'
 
 import { useSortedPaginatedList } from '../../hooks/useSortedPaginatedList'
