+++ frontend/src/scenes/web-analytics/tabs/marketing-analytics/frontend/components/settings/SharedExternalDataSourceConfiguration.tsx
@@ -9,7 +9,7 @@ import { LemonTable } from 'lib/lemon-ui/LemonTable'
 import { cn } from 'lib/utils/css-classes'
 import { DataWarehouseSourceIcon } from 'scenes/data-warehouse/settings/DataWarehouseSourceIcon'
 
-import { SceneSection } from '~/layout/scenes/SceneContent'
+import { SceneSection } from '~/layout/scenes/components/SceneSection'
 import { MARKETING_ANALYTICS_SCHEMA, MarketingAnalyticsColumnsSchemaNames } from '~/queries/schema/schema-general'
 
 import { useSortedPaginatedList } from '../../hooks/useSortedPaginatedList'
