+++ products/revenue_analytics/frontend/settings/FilterTestAccountsConfiguration.tsx
@@ -6,7 +6,7 @@ import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
 import { cn } from 'lib/utils/css-classes'
 import { urls } from 'scenes/urls'
 
-import { SceneSection } from '~/layout/scenes/SceneContent'
+import { SceneSection } from '~/layout/scenes/components/SceneSection'
 
 import { revenueAnalyticsSettingsLogic } from './revenueAnalyticsSettingsLogic'
